Admin - adding logic for expired posts
I've removed the EXPIRED status as we will need to have some sort of periodic job to find
the expired posts and make them expired. Instead, I moved that logic to the database by
creating two object managers:
1. one extending the model with the "not_expired" field
2. another one returning just the posts which can be displayed on the website (published
and not expired).

The first one is used in django admin for marking visually posts which are expired: the list
views have an additional column and the change_form view displayes the "Expired" label for
expired posts.

The second one is used in views - so we don't repeat ourselves too much.